Added different player states that are saved/loaded to/from the player data file + small changes
There is one state for each class (must be in adventure when using classes) and one last state for when a player is not in adventure mode. Every time a player switches state, their inventory, location, health, hunger, level and exp are saved to file for the previous state and loaded for the new state.

Other small changes include adding a check to see if an aspect is unlocked for the opened class before allowing the player to select it, and removing some redundant player tracking variables.